Audit + CleanUp
git-svn-id: https://cchlib.googlecode.com/svn/trunk@8257 bc589049-105d-1923-a3ce-6fcb24364ed1

+++ cchlib-swing/src/main/java/com/googlecode/cchlib/swing/CustomDialogWB.java
@@ -117,7 +117,7 @@ class CustomDialogWB extends JDialog
                 {
                     if( ACTION_CMD_COMPONENT.equals( e.getActionCommand() ) ) {
                         AbstractButton b = AbstractButton.class.cast( e.getSource() );
-                        selectedButtonIndex = Integer.class.cast( b.getClientProperty( CLIENT_PROPERTY_INDEX ) );
+                        selectedButtonIndex = Integer.class.cast( b.getClientProperty( CLIENT_PROPERTY_INDEX ) ).intValue();
 
                         closeDisposeDialog();
                         }
